Create AI
You are blindfolded and 10 coins are placed in front of you on the table. You are allowed to touch the coins but can’t tell which way up they are by feel. You are told that there are 5 coins head up, and 5 coins tails up but not which ones are which. 

Can you make two piles of coins each with the same number of heads up? You can flip the coins any number of times.